fix: dependency cycle in looper resolved
